Consolidate small duplicated helpers across modules
Several small pieces of logic were copy-pasted across unrelated
modules, risking silent drift between copies:

- harmony.rs and thinking.rs each defined a byte-identical overlap()
  helper (longest suffix-of-s/prefix-of-delim match) used by their
  streaming delimiter parsers. Moved to a new src/strutil.rs module
  shared by both.

- shortnames.rs's is_host_component() and verify.rs's inline has_host
  check implemented the same "does this path segment look like a
  registry host" rule, but had drifted: shortnames used
  eq_ignore_ascii_case for the "localhost" match while verify used a
  case-sensitive ==. verify.rs now calls the shared
  shortnames::is_host_component so the two cannot diverge again.

- container.rs's spawn() forwarded two different env-var allowlists
  (GPU_VISIBLE_DEVICE_VARS, LLAMA_CPP_ENV_PASSTHROUGH_VARS) into the
  container args with an identical loop body written out twice.
  Extracted into a forward_env_vars() helper called once per list.

No behavior change other than the case-sensitivity fix in
canonical_repository, which is strictly more permissive (an uppercase
"Localhost" component is now recognized as a host, matching
shortnames' existing behavior) and is not covered by any existing test
expecting the old case-sensitive behavior.
